feat: restructure websocket API gateway imports and add request/response types
