+++ packages/ws-signal/src/create-room.ts
@@ -3,15 +3,15 @@ import {
   APIGatewayProxyWebsocketEventV2,
 } from "aws-lambda";
 
-import { createAPIGatewayEndpoint } from "./api-gateway/endpoint";
-import { createApiGatewayManagementApi } from "./api-gateway/management";
-import { Notifier } from "./api-gateway/notifier";
+import { createAPIGatewayEndpoint } from "./websocket/api-gateway/endpoint";
+import { createApiGatewayManagementApi } from "./websocket/api-gateway/management";
+import { Notifier } from "./websocket/api-gateway/notifier";
 import { createRoomID } from "./core/create-room-id";
 import { DynamoConnections } from "./dynamo-db/dynamo-connections";
 import { createDynamoDBDocument } from "./dynamo-db/dynamo-db-document";
 import { DynamoRooms } from "./dynamo-db/dynamo-rooms";
 import { parseJSON } from "./json/parse";
-import { CreateRoom, CreateRoomSchema } from "./request/create-room";
+import { CreateRoom, CreateRoomSchema } from "./websocket/request/create-room";
 
 /** AWSリージョン */
 const AWS_REGION = process.env.AWS_REGION ?? "";
