+++ packages/ws-signal/src/join-room.ts
@@ -3,13 +3,13 @@ import {
   APIGatewayProxyWebsocketEventV2,
 } from "aws-lambda";
 
-import { createAPIGatewayEndpoint } from "./api-gateway/endpoint";
-import { createApiGatewayManagementApi } from "./api-gateway/management";
-import { Notifier } from "./api-gateway/notifier";
+import { createAPIGatewayEndpoint } from "./websocket/api-gateway/endpoint";
+import { createApiGatewayManagementApi } from "./websocket/api-gateway/management";
+import { Notifier } from "./websocket/api-gateway/notifier";
 import { createDynamoDBDocument } from "./dynamo-db/dynamo-db-document";
 import { DynamoRooms } from "./dynamo-db/dynamo-rooms";
 import { parseJSON } from "./json/parse";
-import { JoinRoomSchema } from "./request/join-room";
+import { JoinRoomSchema } from "./websocket/request/join-room";
 
 /** AWSリージョン */
 const AWS_REGION = process.env.AWS_REGION ?? "";
