+++ packages/ws-signal/src/websocket/lambda-authorizer/authorizer-event.ts
