+++ packages/ws-signal/src/websocket/lambda-authorizer/authorizer-response.ts
