+++ packages/ws-signal/src/websocket/response/error.ts
