+++ packages/ws-signal/src/websocket/response/join-room-rejected.ts
