+++ packages/ws-signal/src/websocket/response/pong.ts
