+++ packages/ws-signal/src/websocket/response/send-guest-signal-accepted.ts
