+++ packages/ws-signal/src/ws-authorizer.ts
@@ -1,10 +1,10 @@
 import { DynamoAuthTokens } from "./dynamo-db/dynamo-auth-tokens";
 import { createDynamoDBDocument } from "./dynamo-db/dynamo-db-document";
-import { AuthorizerEvent } from "./lambda-authorizer/authorizer-event";
+import { AuthorizerEvent } from "./websocket/lambda-authorizer/authorizer-event";
 import {
   failedAuthorize,
   successAuthorize,
-} from "./lambda-authorizer/authorizer-response";
+} from "./websocket/lambda-authorizer/authorizer-response";
 
 /** AWSリージョン */
 const AWS_REGION = process.env.AWS_REGION ?? "";
